+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated documentation in Todo.md to include new examples and tasks.
 - Introduced a new PowerShell function, Add-DiaNodeImage, to generate an HTML table for displaying
   images with customizable properties such as border, style, and image size.
+- Add Join-Hashtable function for merging hashtables
 
 ### Changed
 
@@ -32,6 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add-DiaNodeIcon
   - Add-DiaHTMLNodeTable
 - Updated all workflow files to use windows-latest
+- Update module manifest
+- Enhance Add-DiaHTMLTable with NodeObject support;
+- Refactor Split-Array function name
 
 ### Fixed